support-agent/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
psycopg = {extras = ["binary"], version = "^3.2.3"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
openai = "^1.57.4"
python-dotenv = "^1.0.1"
uvicorn = "^0.33.0"
pydantic = "^2.10.3"
python-multipart = "^0.0.19"
asyncpg = "^0.30.0"
sqlalchemy-utils = "^0.41.2"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
